Fix BMS cell count and pointcloud sor_mean parameter type
- ad_r1m_bringup: Fix NCELLS=3 for 3S BMS, bound cell voltage subindex
- ad_r1m_pointcloud_to_occupancygrid: Remove float cast for sor_mean, int in header

+++ ad_r1m_pointcloud_to_occupancygrid/src/pointcloud_to_grid_node.cpp
@@ -98,7 +98,7 @@ class PointCloudToGrid : public rclcpp::Node {
       } else if (name == "sor_enable") {
         grid_map.filter.sor_enable = param.as_bool();
       } else if (name == "sor_mean") {
-        grid_map.filter.sor_mean = static_cast<float>(param.as_int());
+        grid_map.filter.sor_mean = param.as_int();
       } else if (name == "sor_stddev_mul_thresh") {
         grid_map.filter.sor_stddev_mul_thresh = param.as_double();
       } else if (name == "pass_enable") {
